refactor: move ensure_ibis_table to utils/validation module
