+++ pyretailscience/metrics/base.py
@@ -4,32 +4,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-import ibis
-import pandas as pd
-
 if TYPE_CHECKING:
     import ibis.expr.types as ir
 
 PERCENTAGE_SCALE = 100
-
-
-def ensure_ibis_table(df: pd.DataFrame | ibis.Table) -> ibis.Table:
-    """Convert pandas DataFrame to ibis Table, or validate input is an ibis Table.
-
-    Args:
-        df (pd.DataFrame | ibis.Table): Input data to convert or validate.
-
-    Returns:
-        ibis.Table: An ibis Table representation of the input data.
-
-    Raises:
-        TypeError: If df is neither a pandas DataFrame nor an ibis Table.
-    """
-    if isinstance(df, pd.DataFrame):
-        return ibis.memtable(df)
-    if isinstance(df, ibis.Table):
-        return df
-    raise TypeError("df must be either a pandas DataFrame or an Ibis Table.")
 
 
 def ratio_metric(
