+++ pyretailscience/metrics/distribution/pct_of_stores.py
@@ -14,9 +14,9 @@ from ibis import _
 if TYPE_CHECKING:
     import pandas as pd
 
-from pyretailscience.metrics.base import ensure_ibis_table, ratio_metric
+from pyretailscience.metrics.base import ratio_metric
 from pyretailscience.options import ColumnHelper, get_option
-from pyretailscience.utils.validation import validate_columns
+from pyretailscience.utils.validation import ensure_ibis_table, validate_columns
 
 _TEMP_TOTAL_STORES = "__prs_temp_total_stores__"
